CRM-8768: Dotmailer does not work on PostgreSQL with case sensitivity setting (#19998)
* CRM-8768: Dotmailer does not work on PostgreSQL with case sensitivity setting
- fixed endless export job with big amount of contacts
* CRM-9072: Attempt to remove Address Book Contact from Dotmailer fails integration
* CRM-9073: Export fails when new contact imported by integration and status update
* CRM-8768: Dotmailer does not work on PostgreSQL with case sensitivity setting

+++ ImportExport/Writer/RemovedContactsExportWriter.php
@@ -6,8 +6,6 @@ use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Item\ItemWriterInterface;
 use Akeneo\Bundle\BatchBundle\Step\StepExecutionAwareInterface;
 
-use DotMailer\Api\Exception;
-
 use Psr\Log\LoggerInterface;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
@@ -123,38 +121,36 @@ class RemovedContactsExportWriter implements ItemWriterInterface, StepExecutionA
      */
     protected function removeAddressBookContacts(array $items, EntityRepository $repository, $addressBookOriginId)
     {
+        $removingItemsIds = [];
+        $removingItemsOriginIds = [];
+
+        foreach ($items as $item) {
+            if (empty($item['id']) || empty($item['originId'])) {
+                continue;
+            }
+
+            $removingItemsIds[] = $item['id'];
+            $removingItemsOriginIds[] = $item['originId'];
+            $this->context->incrementDeleteCount();
+        }
+
         try {
-            $removingItemsIds = [];
-            $removingItemsOriginIds = [];
-
-            /**
-             * Remove Dotmailer Contacts from DB.
-             * Smaller, than step batch used because of "IN" max length
-             */
-            foreach ($items as $item) {
-                $removingItemsIds[] = $item['id'];
-                $removingItemsOriginIds[] = $item['originId'];
-
-                $this->context->incrementDeleteCount();
-            }
-
-            if (count($removingItemsIds) > 0) {
-                $this->removeContacts($repository, $removingItemsIds);
-
-                /**
-                 * Remove Dotmailer Contacts from Dotmailer
-                 * Operation is Async in Dotmailer side
-                 */
+            $removingItemsOriginIds = $this->prepareIds($removingItemsOriginIds);
+            if ($removingItemsOriginIds) {
                 $this->transport->removeContactsFromAddressBook($removingItemsOriginIds, $addressBookOriginId);
+                $this->logBatchInfo($removingItemsOriginIds, $addressBookOriginId);
             }
-
-            $this->logBatchInfo($removingItemsOriginIds, $addressBookOriginId);
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             $this->logger
                 ->warning(
-                    "Remove contacts from Address Book '{$addressBookOriginId}' failed. Message: {$e->getMessage()}"
+                    "Remove contacts from Address Book '{$addressBookOriginId}' failed. Message: {$e->getMessage()}",
+                    ['exception' => $e]
                 );
+
+            return;
         }
+
+        $this->removeContacts($repository, $removingItemsIds);
     }
 
     /**
@@ -200,11 +196,32 @@ class RemovedContactsExportWriter implements ItemWriterInterface, StepExecutionA
      */
     protected function removeContacts(EntityRepository $repository, array $removingItemsIds)
     {
+        $removingItemsIds = $this->prepareIds($removingItemsIds);
+        if (!$removingItemsIds) {
+            return;
+        }
+
         $qb = $repository->createQueryBuilder('contact');
         $qb->delete()
             ->where($qb->expr()->in('contact.id', ':removingItemsIds'))
             ->setParameter('removingItemsIds', $removingItemsIds)
             ->getQuery()
             ->execute();
+    }
+
+    /**
+     * @param array $ids
+     * @return array
+     */
+    protected function prepareIds(array $ids)
+    {
+        $ids = array_filter($ids);
+        if (!$ids) {
+            return [];
+        }
+
+        sort($ids);
+
+        return $ids;
     }
 }
